feat: :sparkles: remove deprecated session command and use-case files for cleanup
