--- src/modules/session/application/commands/use-cases/archive-session/archive-session.command.ts
@@ -1,9 +1,0 @@
-import { ICommand } from '@nestjs/cqrs';
-
-export class ArchiveSessionCommand implements ICommand {
-  public readonly uuid: string;
-
-  constructor(root: ArchiveSessionCommand) {
-    this.uuid = root.uuid;
-  }
-}
