--- src/modules/session/application/commands/use-cases/destroy-session/destroy-session.command.ts
@@ -1,9 +1,0 @@
-import { ICommand } from '@nestjs/cqrs';
-
-export class DestroySessionCommand implements ICommand {
-  public readonly uuid: string;
-
-  constructor(root: DestroySessionCommand) {
-    this.uuid = root.uuid;
-  }
-}
