--- src/modules/session/application/queries/get-session/get-session.query.ts
@@ -1,12 +1,0 @@
-import { IQuery } from '@nestjs/cqrs';
-
-export class GetSessionQuery implements IQuery {
-  public readonly uuid: string;
-
-  public readonly withArchived?: boolean;
-
-  constructor(root: GetSessionQuery) {
-    this.uuid = root.uuid;
-    this.withArchived = root.withArchived;
-  }
-}
