+++ src/modules/session/application/use-cases/commands/create-active-session/create-active-session.command.ts
