+++ src/modules/session/application/use-cases/commands/create-invalid-session/create-invalid-session.command.ts
