+++ src/modules/session/application/use-cases/commands/create-invalid-session/create-invalid-session.use-case.ts
@@ -1,24 +1,27 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 
-import { ActivateInvalidSessionCommand } from '@session/application/commands/use-cases/activate-invalid-session/activate-invalid-session.command';
+import { CreateInvalidSessionCommand } from '@session/application/use-cases/commands/create-invalid-session/create-invalid-session.command';
 import { SESSION_REPOSITORY } from '@session/domain/constants/injection-tokens';
 import { ISessionRepositoryContract } from '@session/domain/contracts/session-repository.contract';
-import { ActivateInvalidSessionDomainService } from '@session/domain/domain-services/activate-invalid-session.domain-service';
+import { CreateInvalidSessionDomainService } from '@session/domain/domain-services/create-session-invalid.domain-service';
 import { SessionModel } from '@session/domain/models/session.model';
 
 @Injectable()
-@CommandHandler(ActivateInvalidSessionCommand)
-export class ActiveInvalidSessionUseCase implements ICommandHandler<ActivateInvalidSessionCommand> {
+@CommandHandler(CreateInvalidSessionCommand)
+export class CreateInvalidSessionUseCase implements ICommandHandler<CreateInvalidSessionCommand> {
   constructor(
     @Inject(SESSION_REPOSITORY)
     private readonly repository: ISessionRepositoryContract,
     private readonly publisher: EventPublisher,
-    private readonly activateInvalidSessionDomainService: ActivateInvalidSessionDomainService,
+    private readonly createInvalidSessionDomainService: CreateInvalidSessionDomainService,
   ) {}
 
-  async execute(command: ActivateInvalidSessionCommand): Promise<SessionModel> {
-    const sessionModel = await this.activateInvalidSessionDomainService.go(command);
+  async execute(command: CreateInvalidSessionCommand): Promise<SessionModel> {
+    const sessionModel = await this.createInvalidSessionDomainService.go(
+      command,
+      command.accountUUID,
+    );
 
     const sessionContext = this.publisher.mergeObjectContext(sessionModel);
