--- src/modules/session/application/use-cases/commands/create-session/create-session.use-case.ts
@@ -1,31 +1,0 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
-
-import { CreateSessionCommand } from '@session/application/use-cases/commands/create-session/create-session.command';
-import { SESSION_REPOSITORY } from '@session/domain/constants/injection-tokens';
-import { ISessionRepositoryContract } from '@session/domain/contracts/session-repository.contract';
-import { CreateSessionDomainService } from '@session/domain/domain-services/create-session.domain-service';
-import { SessionModel } from '@session/domain/models/session.model';
-
-@Injectable()
-@CommandHandler(CreateSessionCommand)
-export class CreateSessionUseCase implements ICommandHandler<CreateSessionCommand> {
-  constructor(
-    @Inject(SESSION_REPOSITORY)
-    private readonly repository: ISessionRepositoryContract,
-    private readonly publisher: EventPublisher,
-    private readonly createSessionDomainService: CreateSessionDomainService,
-  ) {}
-
-  async execute(command: CreateSessionCommand): Promise<SessionModel> {
-    const sessionModel = await this.createSessionDomainService.go(command, command.accountUUID);
-
-    const sessionContext = this.publisher.mergeObjectContext(sessionModel);
-
-    await this.repository.persist(sessionModel);
-
-    sessionContext.commit();
-
-    return sessionContext;
-  }
-}
